+++ src/sentry/api/endpoints/organization_group_index.py
@@ -3,12 +3,10 @@ from __future__ import absolute_import, division, print_function
 import functools
 import six
 
-from django.conf import settings
-
 from rest_framework.exceptions import PermissionDenied
 from rest_framework.response import Response
 
-from sentry import features
+from sentry import features, search
 from sentry.api.bases import OrganizationEventsEndpointBase, OrganizationEventPermission
 from sentry.api.helpers.group_index import (
     build_query_params_from_request,
@@ -21,13 +19,11 @@ from sentry.api.serializers import serialize
 from sentry.api.serializers.models.group import StreamGroupSerializerSnuba
 from sentry.api.utils import get_date_range_from_params, InvalidParams
 from sentry.models import Group, GroupStatus
-from sentry.search.snuba.backend import SnubaSearchBackend
+
+
 from sentry.utils.validators import normalize_event_id
 
 ERR_INVALID_STATS_PERIOD = "Invalid stats_period. Valid choices are '', '24h', and '14d'"
-
-
-search = SnubaSearchBackend(**settings.SENTRY_SEARCH_OPTIONS)
 
 
 class OrganizationGroupIndexEndpoint(OrganizationEventsEndpointBase):
@@ -42,6 +38,7 @@ class OrganizationGroupIndexEndpoint(OrganizationEventsEndpointBase):
             query_kwargs.update(extra_query_kwargs)
 
         query_kwargs["environments"] = environments if environments else None
+
         result = search.query(**query_kwargs)
         return result, query_kwargs
