+++ src/sentry/runner/initializer.py
@@ -584,7 +584,8 @@ def validate_snuba():
         return
 
     has_all_snuba_required_backends = (
-        settings.SENTRY_SEARCH == "sentry.search.snuba.SnubaSearchBackend"
+        settings.SENTRY_SEARCH
+        in ("sentry.search.snuba.SnubaSearchBackend", "sentry.utils.services.ServiceDelegator")
         and settings.SENTRY_TAGSTORE == "sentry.tagstore.snuba.SnubaTagStorage"
         and
         # TODO(mattrobenolt): Remove ServiceDelegator check
