+++ src/sentry/search/moresnuba/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import absolute_import, print_function
+
+from .backend import *  # NOQA
